+++ chatglm2/module/attention.py
@@ -1,34 +1,22 @@
 import math
-import copy
-import warnings
-import re
-import sys
-
 import torch
+from torch import nn
 import torch.utils.checkpoint
 import torch.nn.functional as F
-from torch import nn
-from torch.nn import CrossEntropyLoss, LayerNorm
-from torch.nn import CrossEntropyLoss, LayerNorm, MSELoss, BCEWithLogitsLoss
-from torch.nn.utils import skip_init
-from typing import Optional, Tuple, Union, List, Callable, Dict, Any
+from loguru import logger
 
-from transformers.modeling_utils import PreTrainedModel
-from transformers.utils import logging
-from transformers.generation.logits_process import LogitsProcessor
-from transformers.generation.utils import LogitsProcessorList, StoppingCriteriaList, GenerationConfig, ModelOutput
 
 from ..configuration_chatglm import ChatGLMConfig
-from .misc import _config_to_kwargs, split_tensor_along_last_dim
 from .position_embedding import apply_rotary_pos_emb
+from .misc import _config_to_kwargs, split_tensor_along_last_dim
 
 
 class CoreAttention(torch.nn.Module):
-    def __init__(self, config: ChatGLMConfig, layer_number):
+    def __init__(self, config: ChatGLMConfig, layer_number: int, verbose_level='trace'):
         super(CoreAttention, self).__init__()
 
-        self.apply_query_key_layer_scaling = config.apply_query_key_layer_scaling
-        self.attention_softmax_in_fp32 = config.attention_softmax_in_fp32
+        self.apply_query_key_layer_scaling = config.apply_query_key_layer_scaling # True
+        self.attention_softmax_in_fp32 = config.attention_softmax_in_fp32 # True
         if self.apply_query_key_layer_scaling:
             self.attention_softmax_in_fp32 = True
         self.layer_number = max(1, layer_number)
@@ -36,9 +24,9 @@ class CoreAttention(torch.nn.Module):
         projection_size = config.kv_channels * config.num_attention_heads
 
         # Per attention head and per partition values.
-        self.hidden_size_per_partition = projection_size
-        self.hidden_size_per_attention_head = projection_size // config.num_attention_heads
-        self.num_attention_heads_per_partition = config.num_attention_heads
+        self.hidden_size_per_partition = projection_size # 4096
+        self.hidden_size_per_attention_head = projection_size // config.num_attention_heads # 128 = 4096 // 32
+        self.num_attention_heads_per_partition = config.num_attention_heads # 32
 
         coeff = None
         self.norm_factor = math.sqrt(self.hidden_size_per_attention_head)
@@ -48,31 +36,45 @@ class CoreAttention(torch.nn.Module):
         self.coeff = coeff
 
         self.attention_dropout = torch.nn.Dropout(config.attention_dropout)
+        self.verbose_level = verbose_level
 
     def forward(self, query_layer, key_layer, value_layer, attention_mask):
         pytorch_major_version = int(torch.__version__.split('.')[0])
+    
         if pytorch_major_version >= 2:
+            """scaled_dot_product_attention(query, key, value, attn_mask=None, dropout_p=0.0, is_causal=False)"""
+            # [seq_len, bs, 32, 128] --> [bs, 32, seq_len, 128]
             query_layer, key_layer, value_layer = [k.permute(1, 2, 0, 3) for k in [query_layer, key_layer, value_layer]]
+
+            # verbose
+            if self.layer_number == 28:
+                getattr(logger, self.verbose_level)(
+                    f"Q: {list(query_layer.shape)}, KV: {list(key_layer.shape)}, "
+                    f"causal: {attention_mask is None and query_layer.shape[2] == key_layer.shape[2]}, "
+                    f"attention_mask: {attention_mask}")
+            
             if attention_mask is None and query_layer.shape[2] == key_layer.shape[2]:
-                context_layer = torch.nn.functional.scaled_dot_product_attention(query_layer, key_layer, value_layer,
-                                                                                 is_causal=True)
+                # step 1: Causual Attention
+                context_layer = F.scaled_dot_product_attention(query_layer, key_layer, value_layer, is_causal=True)
             else:
+                # step 2 ~ n: `The sequence length of Query is 1`, Q: [bs, 32, 1, 128], K, V: [bs, 32, seq++, 128]
                 if attention_mask is not None:
                     attention_mask = ~attention_mask
-                context_layer = torch.nn.functional.scaled_dot_product_attention(query_layer, key_layer, value_layer,
-                                                                                 attention_mask)
+                context_layer = F.scaled_dot_product_attention(query_layer, key_layer, value_layer, attention_mask) # None (eval mode)
+            
+            # Reshape: [bs, 32, seq_len, 128] --> [seq_len, bs, 32, 128] --> [seq_len, bs, 4096]
             context_layer = context_layer.permute(2, 0, 1, 3)
             new_context_layer_shape = context_layer.size()[:-2] + (self.hidden_size_per_partition,)
             context_layer = context_layer.reshape(*new_context_layer_shape)
         else:
-            # Raw attention scores
+            # TODO: Raw attention scores
 
             # [b, np, sq, sk]
             output_size = (query_layer.size(1), query_layer.size(2), query_layer.size(0), key_layer.size(0))
 
-            # [sq, b, np, hn] -> [sq, b * np, hn]
+            # [sq, b, np, hn] --> [sq, b * np, hn]
             query_layer = query_layer.view(output_size[2], output_size[0] * output_size[1], -1)
-            # [sk, b, np, hn] -> [sk, b * np, hn]
+            # [sk, b, np, hn] --> [sk, b * np, hn]
             key_layer = key_layer.view(output_size[3], output_size[0] * output_size[1], -1)
 
             # preallocting input tensor: [b * np, sq, sk]
@@ -119,7 +121,7 @@ class CoreAttention(torch.nn.Module):
             # Context layer. [sq, b, hp]
             # =========================
 
-            # value_layer -> context layer.
+            # value_layer --> context layer.
             # [sk, b, np, hn] --> [b, np, sq, hn]
 
             # context layer shape: [b, np, sq, hn]
@@ -148,25 +150,25 @@ class SelfAttention(torch.nn.Module):
     and returns output of the same size.
     """
 
-    def __init__(self, config: ChatGLMConfig, layer_number, device=None):
+    def __init__(self, config: ChatGLMConfig, layer_number, device=None, verbose_level='debug'):
         super(SelfAttention, self).__init__()
         self.layer_number = max(1, layer_number)
 
-        self.projection_size = config.kv_channels * config.num_attention_heads
+        self.projection_size = config.kv_channels * config.num_attention_heads # 4096 = 128 * 32
 
         # Per attention head and per partition values.
-        self.hidden_size_per_attention_head = self.projection_size // config.num_attention_heads
+        self.hidden_size_per_attention_head = self.projection_size // config.num_attention_heads # 128 = 4096 // 32
         self.num_attention_heads_per_partition = config.num_attention_heads
 
-        self.multi_query_attention = config.multi_query_attention
-        self.qkv_hidden_size = 3 * self.projection_size
+        self.multi_query_attention = config.multi_query_attention # True
+        self.qkv_hidden_size = 3 * self.projection_size # 12288
         if self.multi_query_attention:
-            self.num_multi_query_groups_per_partition = config.multi_query_group_num
+            self.num_multi_query_groups_per_partition = config.multi_query_group_num # Group Query Attention: 2
             self.qkv_hidden_size = (
                     self.projection_size + 2 * self.hidden_size_per_attention_head * config.multi_query_group_num
-            )
-        self.query_key_value = nn.Linear(config.hidden_size, self.qkv_hidden_size,
-                                         bias=config.add_bias_linear or config.add_qkv_bias,
+            ) # 4608 = (32 + 2 + 2) * num_heads
+        self.query_key_value = nn.Linear(config.hidden_size, self.qkv_hidden_size, # 4096, 4608
+                                         bias=config.add_bias_linear or config.add_qkv_bias, # True
                                          device=device, **_config_to_kwargs(config)
                                          )
 
@@ -182,6 +184,7 @@ class SelfAttention(torch.nn.Module):
             num_attention_heads = self.num_multi_query_groups_per_partition
         else:
             num_attention_heads = self.num_attention_heads_per_partition
+        
         return torch.empty(
             inference_max_sequence_len,
             batch_size,
@@ -191,39 +194,35 @@ class SelfAttention(torch.nn.Module):
             device=device,
         )
 
-    def forward(
-            self, hidden_states, attention_mask, rotary_pos_emb, kv_cache=None, use_cache=True
-    ):
+    def forward(self, hidden_states, attention_mask, rotary_pos_emb, kv_cache=None, use_cache=True):
         # hidden_states: [sq, b, h]
 
         # =================================================
         # Pre-allocate memory for key-values for inference.
         # =================================================
-        # =====================
-        # Query, Key, and Value
-        # =====================
 
-        # Attention heads [sq, b, h] --> [sq, b, (np * 3 * hn)]
+        # Step 1. Attention heads [sq, b, h] --> [sq, b, (32 + 2 + 2) * 128]
         mixed_x_layer = self.query_key_value(hidden_states)
 
+        # Step 2. Split to `QKV`
         if self.multi_query_attention:
             (query_layer, key_layer, value_layer) = mixed_x_layer.split(
                 [
-                    self.num_attention_heads_per_partition * self.hidden_size_per_attention_head,
-                    self.num_multi_query_groups_per_partition * self.hidden_size_per_attention_head,
-                    self.num_multi_query_groups_per_partition * self.hidden_size_per_attention_head,
+                    self.num_attention_heads_per_partition * self.hidden_size_per_attention_head, # 32, 128
+                    self.num_multi_query_groups_per_partition * self.hidden_size_per_attention_head, # 2, 128
+                    self.num_multi_query_groups_per_partition * self.hidden_size_per_attention_head, # 2, 128
                 ],
                 dim=-1,
             )
             query_layer = query_layer.view(
                 query_layer.size()[:-1] + (self.num_attention_heads_per_partition, self.hidden_size_per_attention_head)
-            )
+            ) # [seq_len, bs, 4096] --> [seq_len, bs, `32`, 128]
+            
             key_layer = key_layer.view(
                 key_layer.size()[:-1] + (self.num_multi_query_groups_per_partition, self.hidden_size_per_attention_head)
-            )
+            ) # [seq_len, bs, 256] --> [seq_len, bs, `2`, 128]
             value_layer = value_layer.view(
-                value_layer.size()[:-1]
-                + (self.num_multi_query_groups_per_partition, self.hidden_size_per_attention_head)
+                value_layer.size()[:-1] + (self.num_multi_query_groups_per_partition, self.hidden_size_per_attention_head)
             )
         else:
             new_tensor_shape = mixed_x_layer.size()[:-1] + \
@@ -234,29 +233,31 @@ class SelfAttention(torch.nn.Module):
             # [sq, b, np, 3 * hn] --> 3 [sq, b, np, hn]
             (query_layer, key_layer, value_layer) = split_tensor_along_last_dim(mixed_x_layer, 3)
 
-        # apply relative positional encoding (rotary embedding)
+        # Step 3. apply relative positional encoding (rotary embedding)
         if rotary_pos_emb is not None:
             query_layer = apply_rotary_pos_emb(query_layer, rotary_pos_emb)
             key_layer = apply_rotary_pos_emb(key_layer, rotary_pos_emb)
 
-        # adjust key and value for inference
+        # Step 4. kv_cache: adjust key and value for inference
         if kv_cache is not None:
             cache_k, cache_v = kv_cache
-            key_layer = torch.cat((cache_k, key_layer), dim=0)
+            key_layer = torch.cat((cache_k, key_layer), dim=0) # [seq + 1, bs, 32, 128] <-- [seq, ...] + [1, ...]
             value_layer = torch.cat((cache_v, value_layer), dim=0)
         if use_cache:
-            kv_cache = (key_layer, value_layer)
+            kv_cache = (key_layer, value_layer) if use_cache else None
         else:
             kv_cache = None
 
+        # Step 5. Repeat `key_layer` and `value_layer`
         if self.multi_query_attention:
-            key_layer = key_layer.unsqueeze(-2)
+            key_layer = key_layer.unsqueeze(-2) # [seq_len, bs, 2, 128] --> [seq_len, bs, 2, `1`, 128]
             key_layer = key_layer.expand(
                 -1, -1, -1, self.num_attention_heads_per_partition // self.num_multi_query_groups_per_partition, -1
-            )
+            ) # [seq_len, bs, 2, 1, 128] --> # [seq_len, bs, 2, `16`, 128]
             key_layer = key_layer.contiguous().view(
                 key_layer.size()[:2] + (self.num_attention_heads_per_partition, self.hidden_size_per_attention_head)
-            )
+            ) # [seq_len, bs, 2, 16, 128] --> [seq_len, bs, `32`, 128]
+            
             value_layer = value_layer.unsqueeze(-2)
             value_layer = value_layer.expand(
                 -1, -1, -1, self.num_attention_heads_per_partition // self.num_multi_query_groups_per_partition, -1
@@ -265,16 +266,10 @@ class SelfAttention(torch.nn.Module):
                 value_layer.size()[:2] + (self.num_attention_heads_per_partition, self.hidden_size_per_attention_head)
             )
 
-        # ==================================
-        # core attention computation
-        # ==================================
-
+        # Step 6. core attention computation, 1) query: [seq, 1, 1, ..., 1]
         context_layer = self.core_attention(query_layer, key_layer, value_layer, attention_mask)
 
-        # =================
-        # Output. [sq, b, h]
-        # =================
-
+        # Step 7. Output. [sq, b, h]
         output = self.dense(context_layer)
 
         return output, kv_cache