+++ chatglm2/module/model.py
@@ -1,27 +1,17 @@
-
-
 import torch
-import torch.utils.checkpoint
 from torch import nn
-from torch.nn import LayerNorm
-from torch.nn import LayerNorm
+import torch.utils.checkpoint
 from torch.nn.utils import skip_init
 from typing import Optional, Tuple
+from loguru import logger
 
-from transformers.modeling_outputs import (
-    BaseModelOutputWithPast,
-)
+from transformers.modeling_outputs import BaseModelOutputWithPast
 from transformers.modeling_utils import PreTrainedModel
-from transformers.utils import logging
 
-from ..configuration_chatglm import ChatGLMConfig
-
-from .decoder_layer import GLMBlock
-from .norm import RMSNorm
 from .embedding import Embedding
+from .transformer import GLMTransformer
 from .position_embedding import RotaryEmbedding
-
-logger = logging.get_logger(__name__)
+from ..configuration_chatglm import ChatGLMConfig
 
 
 def default_init(cls, *args, **kwargs):
@@ -58,88 +48,6 @@ class PrefixEncoder(torch.nn.Module):
         else:
             past_key_values = self.embedding(prefix)
         return past_key_values
-
-
-class GLMTransformer(torch.nn.Module):
-    """Transformer class."""
-
-    def __init__(self, config: ChatGLMConfig, device=None):
-        super(GLMTransformer, self).__init__()
-
-        self.fp32_residual_connection = config.fp32_residual_connection
-        self.post_layer_norm = config.post_layer_norm
-
-        # Number of layers.
-        self.num_layers = config.num_layers
-
-        # Transformer layers.
-        def build_layer(layer_number):
-            return GLMBlock(config, layer_number, device=device)
-
-        self.layers = torch.nn.ModuleList([build_layer(i + 1) for i in range(self.num_layers)])
-
-        if self.post_layer_norm:
-            LayerNormFunc = RMSNorm if config.rmsnorm else LayerNorm
-            # Final layer norm before output.
-            self.final_layernorm = LayerNormFunc(config.hidden_size, eps=config.layernorm_epsilon, device=device,
-                                                 dtype=config.torch_dtype)
-
-        self.gradient_checkpointing = False
-
-    def _get_layer(self, layer_number):
-        return self.layers[layer_number]
-
-    def forward(
-            self, hidden_states, attention_mask, rotary_pos_emb, kv_caches=None,
-            use_cache: Optional[bool] = True,
-            output_hidden_states: Optional[bool] = False,
-    ):
-        if not kv_caches:
-            kv_caches = [None for _ in range(self.num_layers)]
-        presents = () if use_cache else None
-        if self.gradient_checkpointing and self.training:
-            if use_cache:
-                logger.warning_once(
-                    "`use_cache=True` is incompatible with gradient checkpointing. Setting `use_cache=False`..."
-                )
-                use_cache = False
-
-        all_self_attentions = None
-        all_hidden_states = () if output_hidden_states else None
-        for index in range(self.num_layers):
-            if output_hidden_states:
-                all_hidden_states = all_hidden_states + (hidden_states,)
-
-            layer = self._get_layer(index)
-            if self.gradient_checkpointing and self.training:
-                layer_ret = torch.utils.checkpoint.checkpoint(
-                    layer,
-                    hidden_states,
-                    attention_mask,
-                    rotary_pos_emb,
-                    kv_caches[index],
-                    use_cache
-                )
-            else:
-                layer_ret = layer(
-                    hidden_states,
-                    attention_mask,
-                    rotary_pos_emb,
-                    kv_cache=kv_caches[index],
-                    use_cache=use_cache
-                )
-            hidden_states, kv_cache = layer_ret
-            if use_cache:
-                presents = presents + (kv_cache,)
-
-        if output_hidden_states:
-            all_hidden_states = all_hidden_states + (hidden_states,)
-
-        # Final layer norm.
-        if self.post_layer_norm:
-            hidden_states = self.final_layernorm(hidden_states)
-
-        return hidden_states, presents, all_hidden_states, all_self_attentions
 
 
 class ChatGLMPreTrainedModel(PreTrainedModel):
@@ -187,31 +95,36 @@ class ChatGLMPreTrainedModel(PreTrainedModel):
 
 
 class ChatGLMModel(ChatGLMPreTrainedModel):
-    def __init__(self, config: ChatGLMConfig, device=None, empty_init=True):
+    def __init__(self, config: ChatGLMConfig, device=None, empty_init=True, verbose_level='debug'):
         super().__init__(config)
         if empty_init:
             init_method = skip_init
         else:
             init_method = default_init
+        
         init_kwargs = {}
         if device is not None:
             init_kwargs["device"] = device
-        self.embedding = init_method(Embedding, config, **init_kwargs)
         self.num_layers = config.num_layers
         self.multi_query_group_num = config.multi_query_group_num
         self.kv_channels = config.kv_channels
+        self.verbose_level = verbose_level
 
         # Rotary positional embeddings
         self.seq_length = config.seq_length
         rotary_dim = (
             config.hidden_size // config.num_attention_heads if config.kv_channels is None else config.kv_channels
-        )
+        ) # 128
 
+        # `module`
+        self.embedding = init_method(Embedding, config, **init_kwargs)
         self.rotary_pos_emb = RotaryEmbedding(rotary_dim // 2, original_impl=config.original_rope, device=device,
                                               dtype=config.torch_dtype)
         self.encoder = init_method(GLMTransformer, config, **init_kwargs)
         self.output_layer = init_method(nn.Linear, config.hidden_size, config.padded_vocab_size, bias=False,
-                                        dtype=config.torch_dtype, **init_kwargs)
+                                        dtype=config.torch_dtype, **init_kwargs) # 在子类 `ChatGLMForConditionalGeneration` 中使用
+        
+        # `pre_seq_len`
         self.pre_seq_len = config.pre_seq_len
         self.prefix_projection = config.prefix_projection
         if self.pre_seq_len is not None:
@@ -239,6 +152,11 @@ class ChatGLMModel(ChatGLMPreTrainedModel):
         past_key_values = past_key_values.permute([2, 1, 0, 3, 4]).split(2)
         return past_key_values
 
+    def quantize(self, weight_bit_width: int):
+        from .quantization import quantize
+        quantize(self.encoder, weight_bit_width)
+        return self
+
     def forward(
             self,
             input_ids,
@@ -251,43 +169,57 @@ class ChatGLMModel(ChatGLMPreTrainedModel):
             output_hidden_states: Optional[bool] = None,
             return_dict: Optional[bool] = None,
     ):
-        output_hidden_states = (
+        # Step 1: check config
+        output_hidden_states = ( # False
             output_hidden_states if output_hidden_states is not None else self.config.output_hidden_states
         )
-        use_cache = use_cache if use_cache is not None else self.config.use_cache
-        return_dict = return_dict if return_dict is not None else self.config.use_return_dict
+        use_cache = use_cache if use_cache is not None else self.config.use_cache # True
+        return_dict = return_dict if return_dict is not None else self.config.use_return_dict # True
 
         batch_size, seq_length = input_ids.shape
 
+        # Step 2: embedding
         if inputs_embeds is None:
-            inputs_embeds = self.embedding(input_ids)
+            inputs_embeds = self.embedding(input_ids) # (seq_len, bs, 4096)
 
-        if self.pre_seq_len is not None:
+        # Step 3: init `pre_seq_len`
+        if self.pre_seq_len is not None: # None
             if past_key_values is None:
                 past_key_values = self.get_prompt(batch_size=batch_size, device=input_ids.device,
                                                   dtype=inputs_embeds.dtype)
             if attention_mask is not None:
-                attention_mask = torch.cat([attention_mask.new_ones((batch_size, self.pre_seq_len)),
-                                            attention_mask], dim=-1)
+                attention_mask = torch.cat(
+                    [attention_mask.new_ones((batch_size, self.pre_seq_len)), attention_mask], dim=-1)
 
+        # Step 4: attention_mask
+        getattr(logger, self.verbose_level)(
+            f"attention_mask: {(list(attention_mask.shape), bool(attention_mask.all())) if attention_mask is not None else 'None'}, "
+            f"full_attention_mask: {list(full_attention_mask.shape) if full_attention_mask is not None else 'None'}"
+        )
         if full_attention_mask is None:
-            if (attention_mask is not None and not attention_mask.all()) or (past_key_values and seq_length != 1):
-                full_attention_mask = self.get_masks(input_ids, past_key_values, padding_mask=attention_mask)
+            # attention_mask = [1] * seq_len
+            if (attention_mask is not None and not attention_mask.all()) or (past_key_values and seq_length != 1): # False, False
+                full_attention_mask = self.get_masks(input_ids, past_key_values, padding_mask=attention_mask) # TODO None
+        if full_attention_mask is not None:
+            getattr(logger, self.verbose_level)(
+                f"full_attention_mask: {list(full_attention_mask.shape) if full_attention_mask is not None else 'None'}"
+            )
 
-        # Rotary positional embeddings
+        # Step 5: Rotary positional embeddings
         rotary_pos_emb = self.rotary_pos_emb(self.seq_length)
         if position_ids is not None:
             rotary_pos_emb = rotary_pos_emb[position_ids]
         else:
             rotary_pos_emb = rotary_pos_emb[None, :seq_length]
-        rotary_pos_emb = rotary_pos_emb.transpose(0, 1).contiguous()
+        rotary_pos_emb = rotary_pos_emb.transpose(0, 1).contiguous() # (seq_len, bs, 32, 2)
 
-        # Run encoder.
+        # Step 6: Run encoder.
         hidden_states, presents, all_hidden_states, all_self_attentions = self.encoder(
             inputs_embeds, full_attention_mask, rotary_pos_emb=rotary_pos_emb,
             kv_caches=past_key_values, use_cache=use_cache, output_hidden_states=output_hidden_states
         )
 
+        # Step 7: Output
         if not return_dict:
             return tuple(v for v in [hidden_states, presents, all_hidden_states, all_self_attentions] if v is not None)
 
@@ -297,9 +229,3 @@ class ChatGLMModel(ChatGLMPreTrainedModel):
             hidden_states=all_hidden_states,
             attentions=all_self_attentions,
         )
-
-    def quantize(self, weight_bit_width: int):
-        from .quantization import quantize
-        quantize(self.encoder, weight_bit_width)
-        return self
-